Add BigO
Previous versions didn't based on BigO

They are based on the step count method of each algorithm. Big O might be better as it is the real time complexity of the algorithm.

We will be comparing the quadratic and linear log algorithms and comparing them to find which algorithm is better.